// packages/evat/src/generated/types.ts
export interface AddressType {
  /**
   * Simple address
   *
   * Egyszerű cím
   */
  simpleAddress?: SimpleAddressType;
  /**
   * Detailed address
   *
   * Részletes cím
   */
  detailedAddress?: DetailedAddressType;
}

/**
 * Agricultural compensation premium
 *
 * Mezőgazdasági kompenzációs felár
 */
export interface AgriculturalCompensationPremiumType {
  /**
   * Count of compensations at 7%
   *
   * 7%-os mértékű kompenzáció darabszáma
   */
  sevenPercentCount?: number;
  /**
   * Count of compensations at 12%
   *
   * 12%-os mértékű kompenzáció darabszáma
   */
  twelvePercentCount?: number;
}
// ...
export interface DeductionOptionsType {
  /**
   * Deduction rate
   *
   * Levonási hányados
   */
  deductionRate?: string;
  /**
   * Deduction amount
   *
   * Levonási összeg
   */
  deductionAmount: string;
}

/**
 * Detailed address data
 *
 * Részletes címadatok
 */
export interface DetailedAddressType {
  /**
   * ISO 3166 alpha-2 country code
   *
   * Az országkód ISO 3166 alpha-2 szabvány szerint
   */
  countryCode: string;
  /**
   * ISO 3166 alpha-2 province code (if appropriate in a given country)
   *
   * Tartomány kódja (amennyiben értelmezhető az adott országban) az ISO 3166-2
   * alpha 2 szabvány szerint
   */
  region?: string;
  /**
   * ZIP code (If can not be interpreted, need to be filled "0000")
   *
   * Irányítószám (amennyiben nem értelmezhető, 0000 értékkel kell kitölteni)
   */
  postalCode: string;
  /**
   * Settlement
   *
   * Település
   */
  city: string;
  /**
   * Name of public place
   *
   * Közterület neve
   */
  streetName: string;
  /**
   * Category of public place
   *
   * Közterület jellege
   */
  publicPlaceCategory: string;
  /**
   * House number
   *
   * Házszám
   */
  number?: string;
  /**
   * Building
   *
   * Épület
   */
  building?: string;
  /**
   * Staircase
   *
   * Lépcsőház
   */
  staircase?: string;
  /**
   * Floor
   *
   * Emelet
   */
  floor?: string;
  /**
   * Door number
   *
   * Ajtó
   */
  door?: string;
  /**
   * Lot number
   *
   * Helyrajzi szám
   */
  lotNumber?: string;
}
// ...
export interface DomesticTaxDataType {
  /**
   * The 8-digit core number section of the tax number
   *
   * Az adószám 8 jegyű törzsszáma
   */
  taxNumber: string;
  /**
   * The 8-digit core number section of the group member's tax number
   *
   * Csoport tag adószámának 8 jegyű törzsszáma
   */
  groupMemberTaxNumber?: string;
}
// ...
export interface GeneralLedgerInformationType {
  /**
   * General ledger account number
   *
   * Főkönyvi számlaszám
   *
   * @remarks At most 2 entries.
   */
  glAccountId: string[];
  /**
   * General ledger posting date
   *
   * Főkönyvi feladás dátuma
   */
  glPostingDate: string;
  /**
   * Unique identifier for general ledger accounting
   *
   * Főkönyvi könyvelés egyedi azonosítója
   */
  glTransactionId: string;
}

/**
 * Indirect customs representative marker type
 *
 * Közvetett vámjogi képviselő jelölő típus
 */
export interface IndirectCustomsRepresentativeType {
  /**
   * Importer taxnumber
   *
   * Importőr adószám
   */
  importerTaxNumber: string;
}
// ...
export interface PartnerInfoType {
  /**
   * Partner's VAT-related status
   *
   * Partner ÁFA szerinti státusza
   */
  partnerStatus: PartnerStatusType;
  /**
   * Partner's tax number
   *
   * Partner adószáma
   */
  partnerTaxData?: PartnerTaxDataType;
  /**
   * Partner's name
   *
   * Partner neve
   */
  partnerName?: string;
  /**
   * Partner's address
   *
   * Partner címadatai
   */
  partnerAddress?: AddressType;
}

/**
 * Partner's VAT-related status
 *
 * Partner áfa szerinti státusza
 */
export type PartnerStatusType = 'NOT_AVAILABLE' | 'PRIVATE_PERSON' | 'DOMESTIC' | 'OTHER';

/**
 * Partner's tax number
 *
 * Partner adószáma
 *
 * @remarks Exactly one of `domesticTaxData`, `communityVatNumber`, `thirdStateTaxId` may be set.
 */
export interface PartnerTaxDataType {
  /**
   * Domestic tax number data
   *
   * Belföldi adószám adatok
   */
  domesticTaxData?: DomesticTaxDataType;
  /**
   * Community tax number
   *
   * Közösségi adószám
   */
  communityVatNumber?: string;
  /**
   * Third state tax identification number
   *
   * Harmadik országbeli adóazonosító
   */
  thirdStateTaxId?: string;
}
// ...
export type PositionTypeType = 'PAYABLE' | 'DEDUCTIBLE' | 'OTHER';
// ...
export interface SimpleAddressType {
  /**
   * ISO 3166 alpha-2 country code
   *
   * Az országkód az ISO 3166 alpha-2 szabvány szerint
   */
  countryCode: string;
  /**
   * ISO 3166 alpha-2 province code (if appropriate in a given country)
   *
   * Tartomány kódja (amennyiben értelmezhető az adott országban) az ISO 3166-2
   * alpha 2 szabvány szerint
   */
  region?: string;
  /**
   * ZIP code (If can not be interpreted, need to be filled "0000")
   *
   * Irányítószám (amennyiben nem értelmezhető, 0000 értékkel kell kitölteni)
   */
  postalCode: string;
  /**
   * Settlement
   *
   * Település
   */
  city: string;
  /**
   * Further address data (name and type of public place, house number, floor,
   * door, lot number, etc.)
   *
   * További címadatok (pl. közterület neve és jellege, házszám, emelet, ajtó,
   * helyrajzi szám, stb.)
   */
  additionalAddressDetail: string;
}
// ...
export type SourceDocumentTypeType =
  'INVOICE' | 'RECEIPT' | 'CUSTOMS_DECLARATION' | 'OTHER' | 'DAILY_SUMMARY';
// ...
export interface TaxInformationType {
  /**
   * Standard tax code of the item
   *
   * A tétel standard adókódja
   */
  standardTaxCode: string;
  /**
   * Corporate tax code of the item
   *
   * A tétel vállalati adókódja
   */
  ownTaxCode?: string;
  /**
   * Indirect customs representative marker
   *
   * Közvetett vámjogi képviselő jelölő
   */
  indirectCustomsRepresentative?: IndirectCustomsRepresentativeType;
  /**
   * Tax position of the item
   *
   * A tétel adózási pozíciója
   *
   * @remarks At most 2 entries.
   */
  taxPosition: TaxPositionType[];
}
// ...
export interface TaxPositionType {
  /**
   * Tax position
   *
   * Adózási pozíció
   */
  positionType: PositionTypeType;
  /**
   * Tax base
   *
   * Adó alapja
   */
  taxBase: string;
  /**
   * Tax amount
   *
   * Adó összege
   */
  taxAmount: string;
  /**
   * Deduction data
   *
   * Levonási adatok
   */
  deductionOptions?: DeductionOptionsType;
}
// ...
export interface VatAnalyticsItemType {
  /**
   * Sequential number of the analytics
   *
   * Analitika sorszáma
   */
  lineNumber: number;
  /**
   * General ledger data
   *
   * Főkönyvi adatok
   */
  generalLedgerInformation?: GeneralLedgerInformationType;
  /**
   * Source document unique identifier
   *
   * Forrás dokumentum egyedi azonosítója
   */
  sourceDocumentId: string;
  /**
   * Source document issue date
   *
   * Forrás dokumentum kiállítási dátuma
   */
  sourceDocumentIssueDate: string;
  /**
   * Type of the source document
   *
   * Forrás dokumentum bizonylat típusa
   */
  sourceDocumentType: SourceDocumentTypeType;
  /**
   * Taxpoint date
   *
   * Adó esedékességi napja
   */
  taxpointDate: string;
  /**
   * Invoice modification or cancellation marker
   *
   * Számla módosítás vagy sztornó jellegű jelölő
   */
  invoiceModificationOrCancellation: boolean;
  /**
   * Partner data
   *
   * Partner adatok
   */
  partnerInfo: PartnerInfoType;
  /**
   * Taxation data
   *
   * Adózási adatok
   */
  taxInformation: TaxInformationType[];
}

/**
 * VAT analytics
 *
 * ÁFA analitika
 */
export interface VatAnalyticsType {
  /**
   * Number of rows in the analytics
   *
   * Az analitika sorainak száma
   */
  totalRowCount: number;
  /**
   * Agricultural compensation premium
   *
   * Mezőgazdasági kompenzációs felár
   */
  agriculturalCompensationPremium?: AgriculturalCompensationPremiumType;
  /**
   * Analytics item data
   *
   * Analitika tételsor adatai
   */
  vatAnalyticsItem?: VatAnalyticsItemType[];
}